@@ -1,17 +1,20 @@
+# frozen_string_literal: true
+
 class RegistrationsController < Devise::RegistrationsController
   private
+
   def sign_up_params
     params.require(:user).permit(:name,
-      :email,
-      :password,
-      :password_confirmation)
+                                 :email,
+                                 :password,
+                                 :password_confirmation)
   end
 
   def account_update_params
     params.require(:user).permit(:name,
-      :email,
-      :password,
-      :password_confirmation,
-      :current_password)
+                                 :email,
+                                 :password,
+                                 :password_confirmation,
+                                 :current_password)
   end
 end
