@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class RelationshipsController < ApplicationController
   before_action :authenticate_user!
 
@@ -5,7 +7,7 @@ class RelationshipsController < ApplicationController
     @user = User.find(params[:followed_id])
     current_user.follow(@user)
     respond_to do |format|
-      format.html{redirect_to @user}
+      format.html { redirect_to @user }
       format.js
     end
   end
@@ -14,7 +16,7 @@ class RelationshipsController < ApplicationController
     @user = Relationship.find(params[:id]).followed
     current_user.unfollow(@user)
     respond_to do |format|
-      format.html{redirect_to @user}
+      format.html { redirect_to @user }
       format.js
     end
   end
