@@ -1,4 +1,6 @@
+# frozen_string_literal: true
+
 class ApplicationMailer < ActionMailer::Base
-  default from: "no-reply@sqrsqr.com"
-  layout "mailer"
+  default from: 'no-reply@sqrsqr.com'
+  layout 'mailer'
 end
