@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: comments
@@ -25,9 +27,9 @@ class Comment < ApplicationRecord
   belongs_to :commentable, polymorphic: true
   has_many :comments, as: :commentable, dependent: :destroy
 
-  default_scope ->{order(created_at: :desc)}
+  default_scope -> { order(created_at: :desc) }
   validates :user_id, presence: true
   validates :commentable_id, presence: true
   validates :commentable_type, presence: true
-  validates :content, presence: true, length: {maximum: 300}
+  validates :content, presence: true, length: { maximum: 300 }
 end
