@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 # == Schema Information
 #
 # Table name: users
@@ -41,32 +43,34 @@ class User < ApplicationRecord
   #   :confirmable, :lockable, :timeoutable, :omniauthable
 
   devise :database_authenticatable, :registerable,
-    :recoverable, :rememberable, :trackable, :validatable,
-    :confirmable, :timeoutable
+         :recoverable, :rememberable, :trackable, :validatable,
+         :confirmable, :timeoutable
 
-  has_many :active_relationships, class_name:  "Relationship",
-    foreign_key: "follower_id",
-    dependent: :destroy
-  has_many :passive_relationships, class_name:  "Relationship",
-    foreign_key: "followed_id",
-    dependent: :destroy
+  has_many :active_relationships, class_name: 'Relationship',
+                                  foreign_key: 'follower_id',
+                                  inverse_of: :follower,
+                                  dependent: :destroy
+  has_many :passive_relationships, class_name: 'Relationship',
+                                   foreign_key: 'followed_id',
+                                   inverse_of: :followed,
+                                   dependent: :destroy
   has_many :following, through: :active_relationships,  source: :followed
   has_many :followers, through: :passive_relationships, source: :follower
 
   has_many :posts, dependent: :destroy
   has_many :comments, dependent: :destroy
 
-  validates :name, presence: true, length: {maximum: 50}
+  validates :name, presence: true, length: { maximum: 50 }
 
-  def follow other_user
+  def follow(other_user)
     following << other_user
   end
 
-  def unfollow other_user
+  def unfollow(other_user)
     following.delete(other_user)
   end
 
-  def following? other_user
+  def following?(other_user)
     following.include?(other_user)
   end
 
