@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
@@ -29,17 +31,17 @@ Rails.application.configure do
   config.assets.compile = false
 
   # Set action mailer default_url
-  config.action_mailer.default_url_options = {host: "localhost", port: 3000}
+  config.action_mailer.default_url_options = { host: 'localhost', port: 3000 }
   config.action_mailer.perform_deliveries = true
   config.action_mailer.raise_delivery_errors = true
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.smtp_settings = {
-    address:              ENV['email_address'],
-    port:                 ENV['email_port'],
-    domain:               ENV['email_domain'],
-    user_name:            ENV['email_username'],
-    password:             ENV['email_password'],
-    authentication:       'plain',
+    address: ENV['email_address'],
+    port: ENV['email_port'],
+    domain: ENV['email_domain'],
+    user_name: ENV['email_username'],
+    password: ENV['email_password'],
+    authentication: 'plain',
     enable_starttls_auto: true
   }
 
@@ -66,7 +68,7 @@ Rails.application.configure do
   config.log_level = :debug
 
   # Prepend all log lines with the following tags.
-  config.log_tags = [ :request_id ]
+  config.log_tags = [:request_id]
 
   # Use a different cache store in production.
   # config.cache_store = :mem_cache_store
@@ -95,7 +97,7 @@ Rails.application.configure do
   # require 'syslog/logger'
   # config.logger = ActiveSupport::TaggedLogging.new(Syslog::Logger.new 'app-name')
 
-  if ENV["RAILS_LOG_TO_STDOUT"].present?
+  if ENV['RAILS_LOG_TO_STDOUT'].present?
     logger           = ActiveSupport::Logger.new(STDOUT)
     logger.formatter = config.log_formatter
     config.logger    = ActiveSupport::TaggedLogging.new(logger)
