@@ -19,7 +19,7 @@ Devise.setup do |config|
   # note that it will be overwritten if you use your own mailer class
   # with default "from" parameter.
   # config.mailer_sender = 'please-change-me-at-config-initializers-devise@example.com'
-  config.mailer_sender = ENV["email_from"]
+  config.mailer_sender = ENV['email_from']
 
   # Configure the class responsible to send e-mails.
   # config.mailer = 'Devise::Mailer'
