@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class CreatePosts < ActiveRecord::Migration[5.2]
   def change
     create_table :posts do |t|
@@ -6,6 +8,6 @@ class CreatePosts < ActiveRecord::Migration[5.2]
 
       t.timestamps
     end
-    add_index :posts, [:user_id, :created_at]
+    add_index :posts, %i[user_id created_at]
   end
 end
