@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class CreateRelationships < ActiveRecord::Migration[5.2]
   def change
     create_table :relationships do |t|
@@ -8,6 +10,6 @@ class CreateRelationships < ActiveRecord::Migration[5.2]
     end
     add_index :relationships, :follower_id
     add_index :relationships, :followed_id
-    add_index :relationships, [:follower_id, :followed_id], unique: true
+    add_index :relationships, %i[follower_id followed_id], unique: true
   end
 end
