@@ -1,22 +1,24 @@
+# frozen_string_literal: true
+
 user = User.new(
-  name:  "123",
-  email: "123@123.com",
-  password: "123123",
-  password_confirmation: "123123",
-  admin: true,
+  name: '123',
+  email: '123@123.com',
+  password: '123123',
+  password_confirmation: '123123',
+  admin: true
 )
 user.skip_confirmation!
 user.save!
 
 30.times do |n|
   name  = Faker::Name.name
-  email = "example-#{n+1}@railstutorial.org"
-  password = "password"
+  email = "example-#{n + 1}@railstutorial.org"
+  password = 'password'
   user = User.new(
-    name:  name,
+    name: name,
     email: email,
     password: password,
-    password_confirmation: password,
+    password_confirmation: password
   )
   user.skip_confirmation!
   user.save!
@@ -25,22 +27,22 @@ end
 users = User.order(:created_at).take(6)
 
 10.times do
-  title = Faker::Lorem.sentence(1)
-  content = [ Faker::Markdown.emphasis,
-              Faker::Markdown.ordered_list,
-              Faker::Markdown.inline_code,
-              Faker::Markdown.block_code,
-              Faker::Markdown.table ].join("\n")
-  users.each { |user| user.posts.create!(title: title, content: content)  }
+  title = Faker::Lorem.sentence(word_count: 1)
+  content = [Faker::Markdown.emphasis,
+             Faker::Markdown.ordered_list,
+             Faker::Markdown.inline_code,
+             Faker::Markdown.block_code,
+             Faker::Markdown.table].join("\n\n")
+  users.each { |u| u.posts.create!(title: title, content: content) }
 end
 users = User.order(:created_at).take(6)
 posts = Post.where(user_id: 1)
 
 10.times do
-  content = Faker::Lorem.sentence(2)
-  users.each do |user|
+  content = Faker::Lorem.sentence(word_count: 2)
+  users.each do |u|
     posts.each do |post|
-      post.comments.create!(content: content, user_id: user.id)
+      post.comments.create!(content: content, user_id: u.id)
     end
   end
 end
@@ -52,4 +54,3 @@ following = users[2..15]
 followers = users[3..20]
 following.each { |followed| user.follow(followed) }
 followers.each { |follower| follower.follow(user) }
-
