@@ -1,11 +1,13 @@
-ENV["RAILS_ENV"] ||= "test"
+# frozen_string_literal: true
 
-require "spec_helper"
-Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each { |f| require f }
-require File.expand_path("../../config/environment", __FILE__)
-abort("The Rails environment is running in production mode!") if Rails.env.production?
+ENV['RAILS_ENV'] ||= 'test'
 
-require "rspec/rails"
+require 'spec_helper'
+Dir[Rails.root.join('spec/support/**/*.rb')].sort.each { |f| require f }
+require File.expand_path('../config/environment', __dir__)
+abort('The Rails environment is running in production mode!') if Rails.env.production?
+
+require 'rspec/rails'
 
 begin
   ActiveRecord::Migration.maintain_test_schema!
